囲みボタンに幅設定を追加
WidthPanelをcomponent化
